requirements.txt
# This file may be used to create an environment using:==
# $ conda create --name <env> --file <this file>==
# platform: osx-arm64==
affine==2.3.0
agate==1.7.1
agate-dbf==0.2.2
agate-excel==0.2.5
agate-sql==0.5.8
anyio==3.5.0
aom==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==22.1.0
aws-c-auth==0.6.26
aws-c-cal==0.5.21
aws-c-common==0.8.14
aws-c-compression==0.2.16
aws-c-event-stream==0.2.20
aws-c-http==0.7.5
aws-c-io==0.13.19
aws-c-mqtt==0.8.6
aws-c-s3==0.2.7
aws-c-sdkutils==0.1.8
aws-checksums==0.1.14
aws-crt-cpp==0.19.8
aws-sdk-cpp==1.10.57
babel==2.9.1
backcall==0.2.0
blas==2.116
blas-devel==3.9.0
bleach==4.1.0
blosc==1.21.3
boost-cpp==1.78.0
brotli==1.0.9
brotli-bin==1.0.9
brunsli==0.1
bzip2==1.0.8
c-ares==1.18.1
c-blosc2==2.8.0
ca-certificates==2023.05.30
cairo==1.16.0
certifi==2023.5.7
cffi==1.15.1
cfitsio==4.2.0
charls==2.4.1
click==8.0.4
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.0.0
comm==0.1.2
contourpy==1.0.5
csvkit==1.0.4
curl==7.88.1
cycler==0.11.0
cytoolz==0.11.0
dask-core==2022.5.0
dav1d==1.0.0
dbfread==2.0.7
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et_xmlfile==1.1.0
executing==0.8.3
exif==1.6.0
exiftool==12.42
expat==2.5.0
ffmpeg==5.1.2
font-ttf-dejavu-sans-mono==2.37
font-ttf-inconsolata==3.000
font-ttf-source-code-pro==2.038
font-ttf-ubuntu==0.83
fontconfig==2.14.2
fonts-conda-ecosystem==1
fonts-conda-forge==1
fonttools==4.25.0
freetype==2.12.1
freexl==1.0.6
fsspec==2022.3.0
gdal==3.6.3
geos==3.11.1
geotiff==1.7.1
gettext==0.21.1
gflags==2.2.2
giflib==5.2.1
glib==2.74.1
glib-tools==2.74.1
glog==0.6.0
gmp==6.2.1
gnutls==3.7.8
graphite2==1.3.13
greenlet==2.0.1
harfbuzz==6.0.0
hdf4==4.2.15
hdf5==1.12.2
icu==70.1
idna==3.4
imagecodecs==2023.1.23
imageio==2.26.0
importlib-metadata==6.0.0
importlib_metadata==6.0.0
importlib_resources==5.2.0
ipykernel==6.19.2
ipython==8.12.0
ipython_genutils==0.2.0
isodate==0.6.0
jasper==4.0.0
jedi==0.18.1
jinja2==3.1.2
json-c==0.16
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==1.13.5
jxrlib==1.1
kealib==1.5.0
kiwisolver==1.4.4
krb5==1.20.1
lame==3.100
lcms2==2.15
leather==0.3.4
lerc==4.0.0
libabseil==20230125.0
libaec==1.0.6
libarrow==10.0.1
libavif==0.11.1
libblas==3.9.0
libbrotlicommon==1.0.9
libbrotlidec==1.0.9
libbrotlienc==1.0.9
libcblas==3.9.0
libcrc32c==1.1.2
libcurl==7.88.1
libcxx==14.0.6
libdap4==3.20.6
libdeflate==1.17
libedit==3.1.20191231
libev==4.33
libevent==2.1.10
libffi==3.4.2
libgdal==3.6.3
libgdal-arrow-parquet==3.6.3
libgfortran==5.0.0
libgfortran5==11.3.0
libglib==2.74.1
libgoogle-cloud==2.8.0
libgrpc==1.52.1
libiconv==1.17
libidn2==2.3.4
libjpeg-turbo==2.1.5.1
libkml==1.3.0
liblapack==3.9.0
liblapacke==3.9.0
libnetcdf==4.9.1
libnghttp2==1.52.0
libopenblas==0.3.21
libopencv==4.7.0
libopus==1.3.1
libpng==1.6.39
libpq==15.2
libprotobuf==3.21.12
librttopo==1.1.0
libsodium==1.0.18
libspatialite==5.0.1
libsqlite==3.40.0
libssh2==1.10.0
libtasn1==4.19.0
libthrift==0.18.1
libtiff==4.5.0
libunistring==0.9.10
libutf8proc==2.8.0
libvpx==1.11.0
libwebp-base==1.3.0
libxcb==1.13
libxml2==2.10.3
libzip==1.9.2
libzlib==1.2.13
libzopfli==1.0.3
llvm-openmp==14.0.6
locket==1.0.0
lz4-c==1.9.4
markupsafe==2.1.1
matplotlib==3.7.1
matplotlib-base==3.7.1
matplotlib-inline==0.1.6
mistune==0.8.4
munkres==1.1.4
nbclassic==0.5.5
nbconvert==5.5.0
nbformat==5.7.0
ncurses==6.3
nest-asyncio==1.5.6
nettle==3.8.1
networkx==2.7.1
notebook==6.5.4
notebook-shim==0.2.2
nspr==4.35
nss==3.89
numpy==1.23.5
numpy-base==1.23.5
olefile==0.46
openblas==0.3.21
opencv==4.7.0
openh264==2.3.1
openjpeg==2.5.0
openpyxl==3.0.9
openssl==3.1.1
orc==1.8.3
p11-kit==0.24.1
packaging==23.0
pandas==1.5.3
pandoc==2.12
pandocfilters==1.5.0
parsedatetime==2.4
parso==0.8.3
partd==1.2.0
pcre2==10.40
perl==5.32.1
pexpect==4.8.0
pickleshare==0.7.5
piexif==1.1.3
pillow==9.5.0
pip==23.0.1
pixman==0.40.0
pkgutil-resolve-name==1.3.10
platformdirs==2.5.2
plum-py==0.8.6
poppler==23.03.0
poppler-data==0.4.12
postgresql==15.2
proj==9.1.1
prometheus_client==0.14.1
prompt-toolkit==3.0.36
psutil==5.9.0
pthread-stubs==0.4
ptyprocess==0.7.0
pure_eval==0.2.2
py-opencv==4.7.0
pycparser==2.21
pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.18.0
python==3.8.16
python-dateutil==2.8.2
python-fastjsonschema==2.16.2
python-slugify==5.0.2
python_abi==3.8
pytimeparse==1.1.8
pytz==2022.7
pywavelets==1.3.0
pyyaml==6.0
pyzmq==25.1.0
qhull==2020.2
rasterio==1.3.6
re2==2023.02.02
readline==8.2
scikit-image==0.19.2
scipy==1.7.3
send2trash==1.8.0
setuptools==65.6.3
six==1.16.0
snappy==1.1.10
sniffio==1.2.0
snuggs==1.4.7
sqlalchemy==1.4.39
sqlite==3.40.0
stack_data==0.2.0
svt-av1==1.4.1
terminado==0.17.1
testpath==0.6.0
text-unidecode==1.3
tifffile==2021.7.2
tiledb==2.13.2
tk==8.6.12
toolz==0.11.2
tornado==6.2
tqdm==4.65.0
traitlets==5.7.1
typing-extensions==4.6.3
typing_extensions==4.6.3
tzcode==2022g
tzdata==2022f
unidecode==1.2.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.37.1
x264==1!164.3095
x265==3.5
xerces-c==3.2.4
xlrd==2.0.1
xorg-libxau==1.0.9
xorg-libxdmcp==1.1.3
xz==5.2.8
yaml==0.2.5
zeromq==4.3.4
zfp==1.0.0
zipp==3.11.0
zlib==1.2.13
zlib-ng==2.0.7
zstd==1.5.2
